+++ application/views/FrontPage.php
@@ -1,1 +1,41 @@
-FRONT PAGE
+<!--- https://startbootstrap.com/template-overviews/small-business/ --->
+<!-- Heading Row -->
+<div class="row">
+	<div class="col-md-8 hidden-sm hidden-xs">
+		<img class="img-responsive img-rounded example-image" src="<?=base_url('assets/img/example.png')?>">
+	</div>
+	<div class="col-md-4">
+		<h1>trackr.moe</h1>
+		<p>This is a site dedicated to tracking manga across multiple aggregate sites.</p>
+		<p>It is intended to make your life easier by consolidating new updates for every series you are reading onto a single page, removing the need to manually check everywhere for new updates.</p>
+		<!--<a class="btn btn-primary btn-lg" href="--><?//=base_url('user/signup')?><!--">Signup!</a>-->
+	</div>
+	<div class="col-md-8 hidden-lg hidden-md">
+		<img class="img-responsive img-rounded example-image" src="<?=base_url('assets/img/example.png')?>">
+	</div>
+</div>
+<!-- /.row -->
+
+<hr>
+
+<!-- Content Row -->
+<div class="row">
+	<div class="col-md-4">
+		<h2>Support for 10+ sites</h2>
+		<p>trackr.moe has support for 10+ sites, including bato.to, webtoons.com, MangaStream, MangaPanda and <a href="<?=base_url('help')?>">more</a>.</p>
+		<p>Support for new sites are added often and we also take requests for new sites to support!</p>
+	</div>
+	<!-- /.col-md-4 -->
+	<div class="col-md-4">
+		<h2>Easy to use</h2>
+		<p>Simply signup, install the userscript and you're ready to go. No fuss!</p>
+	</div>
+	<!-- /.col-md-4 -->
+	<div class="col-md-4">
+		<h2>Open Source!</h2>
+		<p>trackr.moe is open source!</p>
+		<p>You can find us on <a href="https://github.com/DakuTree/manga-tracker">Github</a>!</p>
+	</div>
+	<!-- /.col-md-4 -->
+</div>
+<!-- /.row -->